android: move the MediaFormat bindings to the androidjni library
