+++ sys/androidmedia/gstamc.c
@@ -53,9 +53,6 @@ static gboolean ignore_unknown_color_formats = FALSE;
 
 static gboolean accepted_color_formats (GstAmcCodecType * type,
     gboolean is_encoder);
-static gboolean
-gst_amc_format_get_jstring (GstAmcFormat * format, const gchar * key,
-    jstring * value);
 
 /* Global cached references */
 static struct
@@ -109,23 +106,7 @@ static struct
   jfieldID presentation_time_us;
   jfieldID size;
 } media_codec_buffer_info;
-static struct
-{
-  jclass klass;
-  jmethodID create_audio_format;
-  jmethodID create_video_format;
-  jmethodID to_string;
-  jmethodID contains_key;
-  jmethodID get_float;
-  jmethodID set_float;
-  jmethodID get_integer;
-  jmethodID set_integer;
-  jmethodID get_string;
-  jmethodID set_string;
-  jmethodID get_byte_buffer;
-  jmethodID set_byte_buffer;
-  jmethodID set_feature_enabled;
-} media_format;
+
 static struct
 {
   jclass klass;
@@ -590,7 +571,6 @@ error:
   return crypto_ctx->mdrm && crypto_ctx->mcrypto && crypto_ctx->mdrm_session_id;
 }
 
-
 GstAmcCodec *
 gst_amc_codec_new (const gchar * name)
 {
@@ -727,9 +707,6 @@ gst_amc_codec_enable_adaptive_playback (GstAmcCodec * codec,
 {
   gint width = 0;
   gint height = 0;
-  jstring jtmpstr;
-
-  JNIEnv *env = gst_jni_get_env ();
 
   if (!gst_amc_codec_is_feature_supported (codec, format, "adaptive-playback") ||
       !gst_amc_format_get_int (format, "width", &width) ||
@@ -737,16 +714,13 @@ gst_amc_codec_enable_adaptive_playback (GstAmcCodec * codec,
     return FALSE;
   }
 
-  jtmpstr = (*env)->NewStringUTF (env, "adaptive-playback");
-  J_CALL_VOID (format->object, media_format.set_feature_enabled, jtmpstr, 1);
+  gst_amc_format_set_feature_enabled (format, "adaptive-playback", TRUE);
 
   GST_DEBUG ("Setting max-width = %d max-height = %d", width, height);
   gst_amc_format_set_int (format, "max-width", width);
   gst_amc_format_set_int (format, "max-height", height);
   gst_amc_format_set_int (format, "adaptive-playback", 1);
 
-error:
-  J_DELETE_LOCAL_REF (jtmpstr);
   return TRUE;
 }
 
@@ -754,19 +728,13 @@ gboolean
 gst_amc_codec_enable_tunneled_video_playback (GstAmcCodec * codec,
     GstAmcFormat * format, gint audio_session_id)
 {
-  jstring jtmpstr;
-  JNIEnv *env = gst_jni_get_env ();
-
-  if (!gst_amc_codec_is_feature_supported (codec, format, "tunneled-playback")) {
+  if (!gst_amc_codec_is_feature_supported (codec, format, GST_AMC_MEDIA_FORMAT_TUNNELED_PLAYBACK)) {
     return FALSE;
   }
 
-  jtmpstr = (*env)->NewStringUTF (env, "tunneled-playback");
-  J_CALL_VOID (format->object, media_format.set_feature_enabled, jtmpstr, 1);
+  gst_amc_format_set_feature_enabled (format, GST_AMC_MEDIA_FORMAT_TUNNELED_PLAYBACK, TRUE);
   gst_amc_format_set_int (format, "audio-session-id", audio_session_id);
 
-error:
-  J_DELETE_LOCAL_REF (jtmpstr);
   return TRUE;
 }
 
@@ -1172,313 +1140,7 @@ error:
   return ret;
 }
 
-GstAmcFormat *
-gst_amc_format_new_audio (const gchar * mime, gint sample_rate, gint channels)
-{
-  GstAmcFormat *format = NULL;
-  jstring mime_str = NULL;
-  jobject object = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (mime);
 
-  mime_str = (*env)->NewStringUTF (env, mime);
-  if (mime_str == NULL)
-    goto error;
-
-  format = g_slice_new0 (GstAmcFormat);
-  J_CALL_STATIC_OBJ (object /* = */ , media_format, create_audio_format,
-      mime_str, sample_rate, channels);
-  AMC_CHK (object);
-
-  format->object = (*env)->NewGlobalRef (env, object);
-  AMC_CHK (format->object);
-
-done:
-  J_DELETE_LOCAL_REF (object);
-  J_DELETE_LOCAL_REF (mime_str);
-  return format;
-
-error:
-  GST_ERROR ("Failed to create format '%s'", mime ? mime : "NULL");
-  if (format)
-    g_slice_free (GstAmcFormat, format);
-  format = NULL;
-  goto done;
-}
-
-GstAmcFormat *
-gst_amc_format_new_video (const gchar * mime, gint width, gint height)
-{
-  GstAmcFormat *format = NULL;
-  jstring mime_str = NULL;
-  jobject object = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (mime && width && height);
-
-  mime_str = (*env)->NewStringUTF (env, mime);
-  if (mime_str == NULL)
-    goto error;
-
-  format = g_slice_new0 (GstAmcFormat);
-
-  J_CALL_STATIC_OBJ (object /* = */ , media_format, create_video_format,
-      mime_str, width, height);
-
-  AMC_CHK (object);
-  format->object = (*env)->NewGlobalRef (env, object);
-  AMC_CHK (format->object);
-
-done:
-  J_DELETE_LOCAL_REF (object);
-  J_DELETE_LOCAL_REF (mime_str);
-  return format;
-
-error:
-  GST_ERROR ("Failed to create format '%s',"
-      " width = %d, height = %d", mime ? mime : "NULL", width, height);
-  if (format)
-    g_slice_free (GstAmcFormat, format);
-  format = NULL;
-  goto done;
-}
-
-void
-gst_amc_format_free (GstAmcFormat * format)
-{
-  JNIEnv *env = gst_jni_get_env ();
-  g_return_if_fail (format != NULL);
-
-  J_DELETE_GLOBAL_REF (format->object);
-  g_slice_free (GstAmcFormat, format);
-}
-
-
-gchar *
-gst_amc_format_to_string (GstAmcFormat * format)
-{
-  jstring v_str = NULL;
-  gchar *ret = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format);
-
-  J_CALL_OBJ (v_str /* = */ , format->object, media_format.to_string);
-  ret = gst_amc_get_string_utf8 (env, v_str);
-error:
-  J_DELETE_LOCAL_REF (v_str);
-  return ret;
-}
-
-gboolean
-gst_amc_format_contains_key (GstAmcFormat * format, const gchar * key)
-{
-  gboolean ret = FALSE;
-  jstring key_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format && key);
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  J_CALL_BOOL (ret /* = */ , format->object, media_format.contains_key,
-      key_str);
-error:
-  J_DELETE_LOCAL_REF (key_str);
-  return ret;
-}
-
-gboolean
-gst_amc_format_get_float (GstAmcFormat * format, const gchar * key,
-    gfloat * value)
-{
-  gboolean ret = FALSE;
-  jstring key_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format && key && value);
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  J_CALL_FLOAT (*value /* = */ , format->object, media_format.get_float,
-      key_str);
-
-  ret = TRUE;
-error:
-  J_DELETE_LOCAL_REF (key_str);
-  return ret;
-}
-
-void
-gst_amc_format_set_float (GstAmcFormat * format, const gchar * key,
-    gfloat value)
-{
-  jstring key_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format && key);
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  J_CALL_VOID (format->object, media_format.set_float, key_str, value);
-error:
-  J_DELETE_LOCAL_REF (key_str);
-}
-
-gboolean
-gst_amc_format_get_int (const GstAmcFormat * format, const gchar * key,
-    gint * value)
-{
-  gboolean ret = FALSE;
-  jstring key_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format && key && value);
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  J_CALL_INT (*value /* = */ , format->object, media_format.get_integer,
-      key_str);
-
-  ret = TRUE;
-error:
-  J_DELETE_LOCAL_REF (key_str);
-  return ret;
-}
-
-void
-gst_amc_format_set_int (GstAmcFormat * format, const gchar * key, gint value)
-{
-  jstring key_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format && key);
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  J_CALL_VOID (format->object, media_format.set_integer, key_str, value);
-error:
-  J_DELETE_LOCAL_REF (key_str);
-}
-
-static gboolean
-gst_amc_format_get_jstring (GstAmcFormat * format, const gchar * key,
-    jstring * value)
-{
-  gboolean ret = FALSE;
-  jstring key_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-
-  AMC_CHK (format && key && value);
-  *value = NULL;
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  J_CALL_OBJ (*value /* = */ , format->object, media_format.get_string,
-      key_str);
-
-  ret = TRUE;
-error:
-  J_DELETE_LOCAL_REF (key_str);
-  return ret;
-}
-
-gboolean
-gst_amc_format_get_string (GstAmcFormat * format, const gchar * key,
-    gchar ** value)
-{
-  gboolean ret = FALSE;
-  jstring tmp_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-
-  if (!gst_amc_format_get_jstring (format, key, &tmp_str))
-    goto error;
-
-  *value = gst_amc_get_string_utf8 (env, tmp_str);
-  AMC_CHK (*value);
-  ret = TRUE;
-error:
-  J_DELETE_LOCAL_REF (tmp_str);
-  return ret;
-}
-
-void
-gst_amc_format_set_string (GstAmcFormat * format, const gchar * key,
-    const gchar * value)
-{
-  jstring key_str = NULL;
-  jstring v_str = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-
-  AMC_CHK (format && key && value);
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  v_str = (*env)->NewStringUTF (env, value);
-  AMC_CHK (v_str);
-
-  J_CALL_VOID (format->object, media_format.set_string, key_str, v_str);
-error:
-  J_DELETE_LOCAL_REF (key_str);
-  J_DELETE_LOCAL_REF (v_str);
-}
-
-gboolean
-gst_amc_format_get_buffer (GstAmcFormat * format, const gchar * key,
-    GstBuffer ** value)
-{
-  gboolean ret = FALSE;
-  jstring key_str = NULL;
-  jobject v = NULL;
-  guint8 *data;
-  gsize size;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format && key && value);
-
-  *value = NULL;
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  J_CALL_OBJ (v /* = */ , format->object, media_format.get_byte_buffer,
-      key_str);
-
-  data = (*env)->GetDirectBufferAddress (env, v);
-  AMC_CHK (data);
-
-  size = (*env)->GetDirectBufferCapacity (env, v);
-  *value = gst_buffer_new_and_alloc (size);
-  memcpy (GST_BUFFER_DATA (*value), data, size);
-
-  ret = TRUE;
-error:
-  J_DELETE_LOCAL_REF (key_str);
-  J_DELETE_LOCAL_REF (v);
-  return ret;
-}
-
-void
-gst_amc_format_set_buffer (GstAmcFormat * format, const gchar * key,
-    GstBuffer * value)
-{
-  jstring key_str = NULL;
-  jobject v = NULL;
-  JNIEnv *env = gst_jni_get_env ();
-  AMC_CHK (format && key && value);
-
-  key_str = (*env)->NewStringUTF (env, key);
-  AMC_CHK (key_str);
-
-  /* FIXME: The buffer must remain valid until the codec is stopped */
-  v = (*env)->NewDirectByteBuffer (env, GST_BUFFER_DATA (value),
-      GST_BUFFER_SIZE (value));
-  AMC_CHK (v);
-
-  J_CALL_VOID (format->object, media_format.set_byte_buffer, key_str, v);
-error:
-  J_DELETE_LOCAL_REF (key_str);
-  J_DELETE_LOCAL_REF (v);
-}
 
 
 static jclass
@@ -1504,6 +1166,8 @@ get_java_classes (void)
   jclass tmp;
 
   GST_DEBUG ("Retrieving Java classes");
+
+  gst_amc_media_format_init ();
 
   env = gst_jni_get_env ();
 
@@ -1654,72 +1318,7 @@ get_java_classes (void)
     goto done;
   }
 
-  tmp = (*env)->FindClass (env, "android/media/MediaFormat");
-  if (!tmp) {
-    ret = FALSE;
-    (*env)->ExceptionClear (env);
-    GST_ERROR ("Failed to get format class");
-    goto done;
-  }
-  media_format.klass = (*env)->NewGlobalRef (env, tmp);
-  if (!media_format.klass) {
-    ret = FALSE;
-    (*env)->ExceptionClear (env);
-    GST_ERROR ("Failed to get format class global reference");
-    goto done;
-  }
 
-  media_format.create_audio_format =
-      (*env)->GetStaticMethodID (env, media_format.klass, "createAudioFormat",
-      "(Ljava/lang/String;II)Landroid/media/MediaFormat;");
-  media_format.create_video_format =
-      (*env)->GetStaticMethodID (env, media_format.klass, "createVideoFormat",
-      "(Ljava/lang/String;II)Landroid/media/MediaFormat;");
-  media_format.to_string =
-      (*env)->GetMethodID (env, media_format.klass, "toString",
-      "()Ljava/lang/String;");
-  media_format.contains_key =
-      (*env)->GetMethodID (env, media_format.klass, "containsKey",
-      "(Ljava/lang/String;)Z");
-  media_format.get_float =
-      (*env)->GetMethodID (env, media_format.klass, "getFloat",
-      "(Ljava/lang/String;)F");
-  media_format.set_float =
-      (*env)->GetMethodID (env, media_format.klass, "setFloat",
-      "(Ljava/lang/String;F)V");
-  media_format.get_integer =
-      (*env)->GetMethodID (env, media_format.klass, "getInteger",
-      "(Ljava/lang/String;)I");
-  media_format.set_integer =
-      (*env)->GetMethodID (env, media_format.klass, "setInteger",
-      "(Ljava/lang/String;I)V");
-  media_format.get_string =
-      (*env)->GetMethodID (env, media_format.klass, "getString",
-      "(Ljava/lang/String;)Ljava/lang/String;");
-  media_format.set_string =
-      (*env)->GetMethodID (env, media_format.klass, "setString",
-      "(Ljava/lang/String;Ljava/lang/String;)V");
-  media_format.get_byte_buffer =
-      (*env)->GetMethodID (env, media_format.klass, "getByteBuffer",
-      "(Ljava/lang/String;)Ljava/nio/ByteBuffer;");
-  media_format.set_byte_buffer =
-      (*env)->GetMethodID (env, media_format.klass, "setByteBuffer",
-      "(Ljava/lang/String;Ljava/nio/ByteBuffer;)V");
-  media_format.set_feature_enabled =
-      (*env)->GetMethodID (env, media_format.klass, "setFeatureEnabled",
-      "(Ljava/lang/String;Z)V");
-
-  if (!media_format.create_audio_format || !media_format.create_video_format
-      || !media_format.contains_key || !media_format.get_float
-      || !media_format.set_float || !media_format.get_integer
-      || !media_format.set_integer || !media_format.get_string
-      || !media_format.set_string || !media_format.get_byte_buffer
-      || !media_format.set_byte_buffer || !media_format.set_feature_enabled) {
-    ret = FALSE;
-    (*env)->ExceptionClear (env);
-    GST_ERROR ("Failed to get format methods");
-    goto done;
-  }
 
   /* MediaCodecInfo */
   media_codec_info.klass = j_find_class (env, "android/media/MediaCodecInfo");
